Implement clean_food_data function.
clean_food_data() makes sure that food_info categories with values that
do not match the expected type are set to default values of 1.0. Along
with implementing clean_food_data, a lot of error handling for
TypeErrors and ValueErrors in the rest of the view functions was able to
be removed.